requirements.txt
Django==1.10.1
djangorestframework==3.4.6
django-filter==0.14.0
django-import-export==0.5.0
django-rest-swagger==0.3.0
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
